Fix for reviews
