+++ configs/base.yaml
@@ -63,8 +63,6 @@ dur_loss: mse # huber|mol
 ###########
 # train and eval
 ###########
-save_ckpt: true
-save_last: false
 num_ckpt_keep: 3
 accumulate_grad_batches: 1
 log_interval: 100
